+++ dvha/models/regression.py
@@ -17,6 +17,7 @@ from dvha.models.random_forest import RandomForestFrame
 from dvha.dialogs.export import save_data_to_file
 from dvha.dialogs.main import SelectRegressionVariablesDialog
 from dvha.paths import ICONS, MODELS_DIR
+from dvha.tools.machine_learning import get_gradient_boosting
 from dvha.tools.utilities import set_msw_background_color, get_tree_ctrl_image, get_window_size
 
 
@@ -390,10 +391,10 @@ class MultiVarResultsFrame(wx.Frame):
                'regression': self.plot.reg}
         pub.sendMessage('control_chart_set_model', **msg)
 
-        algorithms = ['Random Forest', 'Support Vector Machines', 'Decision Trees', 'Gradient Boosted']
+        algorithms = ['Random Forest', 'Support Vector Machines', 'Decision Trees', 'Gradient Boosting']
         self.button = {key: wx.Button(self, wx.ID_ANY, key) for key in algorithms}
         for key in algorithms:
-            if key not in {'Random Forest'}:
+            if key not in {'Random Forest', 'Gradient Boosting'}:
                 self.button[key].Disable()
 
         self.button_export = wx.Button(self, wx.ID_ANY, 'Export Plot Data')
@@ -409,6 +410,7 @@ class MultiVarResultsFrame(wx.Frame):
 
     def __do_bind(self):
         self.Bind(wx.EVT_BUTTON, self.on_random_forest, id=self.button['Random Forest'].GetId())
+        self.Bind(wx.EVT_BUTTON, self.on_gradient_boosting, id=self.button['Gradient Boosting'].GetId())
         self.Bind(wx.EVT_BUTTON, self.on_export, id=self.button_export.GetId())
         self.Bind(wx.EVT_BUTTON, self.on_save_plot, id=self.button_save_plot.GetId())
         self.Bind(wx.EVT_BUTTON, self.on_save_model, id=self.button_save_model.GetId())
@@ -441,6 +443,9 @@ class MultiVarResultsFrame(wx.Frame):
 
     def on_random_forest(self, evt):
         RandomForestFrame(**self.plot.final_stats_data)
+
+    def on_gradient_boosting(self, evt):
+        RandomForestFrame(title='Gradient Boosting', regressor=get_gradient_boosting, **self.plot.final_stats_data)
 
     def on_export(self, evt):
         save_data_to_file(self, 'Save multi-variable regression data to csv', self.plot.get_csv_data())